requirements.txt
# Core Dependencies
openai==1.12.0
anthropic==0.79.0
python-dotenv==1.0.0

# RAG & Vector DB
faiss-cpu==1.7.4
numpy==1.24.3
pypdf==4.0.1

# Route Analysis
pandas==2.0.3
scipy==1.10.1

# Data Integration
requests==2.31.0

# Utilities
tiktoken==0.5.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3
